Add relations endpoint
