+++ transmart-core-db/grails-app/services/org/transmartproject/db/ontology/MDStudiesService.groovy
@@ -26,6 +26,8 @@ import java.util.concurrent.ConcurrentHashMap
 import java.util.concurrent.ConcurrentSkipListMap
 import java.util.stream.Collectors
 
+import static org.transmartproject.core.users.AuthorisationHelper.getPUBLIC_TOKENS
+
 @Transactional
 @CompileStatic
 class MDStudiesService implements MDStudiesResource, ApplicationRunner {
@@ -153,6 +155,15 @@ class MDStudiesService implements MDStudiesResource, ApplicationRunner {
             throw new NoSuchResourceException("No study found for specified studyIds.")
         }
         studiesForUser
+    }
+
+    Boolean areAllStudiesPublic() {
+        def criteria = DetachedCriteria.forClass(Study)
+        criteria.add(Restrictions.not(Restrictions.in('secureObjectToken', PUBLIC_TOKENS)))
+        List<MDStudy> privateStudies = criteria.getExecutableCriteria(sessionFactory.currentSession)
+                .list().stream()
+                .collect(Collectors.toList())
+        privateStudies.size() == 0
     }
 
     private void checkAccessToStudy(Study study, User user, id) {
